rename test projects to match new convention
- add .editorconfig to have underscore prefix for private
- add .vimspector.json to have debug info
