+++ .gitignore
@@ -1,5 +1,4 @@
 # vscode files
-.editorconfig
 .vscode
 .vs
