+++ .vimspector.json
@@ -1,0 +1,18 @@
+ {
+   "configurations": {
+     "Do.Test.Blueprints.Service": {
+       "adapter": "netcoredbg",
+       "configuration": {
+         "request": "launch",
+         "program": "${workspaceRoot}/test/blueprints/Do.Test.Blueprints.Service/bin/Debug/net6.0/Do.Test.Blueprints.Service.dll",
+         "cwd": "${workspaceRoot}/test/blueprints/Do.Test.Blueprints.Service/bin/Debug/net6.0",
+         "args": [],
+         "stopAtEntry": false,
+         "env": {
+             "ASPNETCORE_ENVIRONMENT": "Development",
+             "ASPNETCORE_URLS": "https://localhost:7254;http://localhost:5151"
+         }
+       }
+     }
+   }
+ }
