+++ Do.sln
@@ -13,7 +13,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{C407EF64-9
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "blueprints", "blueprints", "{F1F041D0-E073-455C-96F9-233FDF5CC2A8}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Do.Blueprints.Service.Test", "test\blueprints\Do.Blueprints.Service.Test\Do.Blueprints.Service.Test.csproj", "{F8B3EFE9-7BC4-48C3-98D8-97500A2486C6}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Do.Test.Blueprints.Service", "test\blueprints\Do.Test.Blueprints.Service\Do.Test.Blueprints.Service.csproj", "{F8B3EFE9-7BC4-48C3-98D8-97500A2486C6}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "core", "core", "{FC5EA7BE-2E36-4CC7-8CE0-DD8D69F8CF85}"
 EndProject
@@ -21,7 +21,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Do.Architecture", "src\core
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "core", "core", "{BE392BCA-BDD7-43F6-BF11-9AC259F82CE6}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Do.Architecture.Test", "test\core\Do.Architecture.Test\Do.Architecture.Test.csproj", "{45D92BDD-5EA8-4B6C-A6F0-38C9DE139A32}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Do.Test.Architecture", "test\core\Do.Test.Architecture\Do.Test.Architecture.csproj", "{45D92BDD-5EA8-4B6C-A6F0-38C9DE139A32}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
