+++ src/core/Do.Architecture/Branding/IBanner.cs
@@ -2,5 +2,5 @@
 
 public interface IBanner
 {
-   void Print();
+    void Print();
 }
