--- test/core/Do.Architecture.Test/ArchitectureStubberExtensions.cs
@@ -1,15 +1,0 @@
-using Do.Branding;
-
-namespace Do.Test;
-
-public static class ArchitectureStubberExtensions
-{
-    public static Build ABuild(this Spec.Stubber source,
-        IBanner? banner = default
-    )
-    {
-        banner ??= new Mock<IBanner>().Object;
-
-        return new(banner);
-    }
-}
