+++ test/core/Do.Test.Architecture/Do.Test.Architecture.csproj
