+++ test/core/Do.Test.Architecture/Spec.cs
